requirements.txt
pytesseract==0.3.7
requests==2.22.0
pdf2image==1.14.0
numpy==1.15.3
Django==3.1.7
djangorestframework==3.12.4
face_recognition==1.3.0
Pillow==8.2.0
python_Levenshtein==0.12.2
